pyproject.toml
[tool.poetry]
name = "remaster"
version = "0.1.0"
description = ""
authors = ["Arthur Koehl <avkoehl@ucdavis.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rioxarray = "^0.17.0"
numba = "^0.60.0"
scipy = "^1.15.2"
numpy = "^2.0.0"
geopandas = "^1.0.1"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
rasterio = "^1.4.3"
valleyx = { git = "ssh://git@github.com/avkoehl/valleyx.git", branch = "main" }
rem = { git = "ssh://git@github.com/avkoehl/rem.git", branch = "main" }
matplotlib = "^3.10.1"
ruptures = "^1.1.9"
xarray-spatial = "^0.4.0"
shapelysmooth = "^0.2.1"
whitebox-workflows = "^1.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
